PageEditDevMode file; conditional option display
