+++ src/core/page/index.ts
@@ -2,3 +2,4 @@ export * from './NewPage';
 export * from './PageEdit';
 export * from './PageDelete';
 export * from './NewArticlePage';
+export * from './PageEditDevMode';
